Add env-file secrets, add postgres backend, add docker compose for postgres.

+++ users/models.py
@@ -3,7 +3,7 @@ from django.contrib.auth.models import AbstractUser
 
 
 class User(AbstractUser):
-    uid = models.IntegerField(
+    uid = models.BigAutoField(
         primary_key=True,
         verbose_name='user_id'
     )
